pyproject.toml
[project]
name = "lang-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "black>=25.9.0",
    "ipython>=9.6.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langchain-pinecone>=0.2.12",
    "langchain-qdrant>=0.2.1",
    "langchain-tavily>=0.2.12",
    "langchainhub>=0.1.21",
    "langgraph>=0.6.10",
    "pyright>=1.1.406",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]
